requirements.txt
apify-client>=1.7.0
anthropic>=0.25.0
openai>=1.30.0
yt-dlp>=2024.5.27
python-dotenv>=1.0.0
gspread>=6.0.0
google-auth>=2.29.0
pydantic>=2.7.0
rich>=13.7.0
typer>=0.12.0
httpx>=0.27.0
